Updated all commands, the plugin now compiles but is still missing textactions, pagination, econ and placeholders

+++ src/main/java/com/hiroku/tournaments/commands/PresetsCommand.java
@@ -1,120 +1,120 @@
 package com.hiroku.tournaments.commands;
 
 import com.google.common.collect.ImmutableMap;
+import com.happyzleaf.tournaments.Text;
+import com.happyzleaf.tournaments.User;
+import com.happyzleaf.tournaments.args.ChoiceSetArgument;
 import com.hiroku.tournaments.Presets;
 import com.hiroku.tournaments.Zones;
 import com.hiroku.tournaments.api.Preset;
 import com.hiroku.tournaments.api.Tournament;
 import com.hiroku.tournaments.obj.Zone;
-import org.spongepowered.api.command.CommandException;
-import org.spongepowered.api.command.CommandResult;
-import org.spongepowered.api.command.CommandSource;
-import org.spongepowered.api.command.args.CommandContext;
-import org.spongepowered.api.command.args.GenericArguments;
-import org.spongepowered.api.command.spec.CommandExecutor;
-import org.spongepowered.api.command.spec.CommandSpec;
-import org.spongepowered.api.entity.living.player.Player;
-import org.spongepowered.api.text.Text;
-import org.spongepowered.api.text.action.TextActions;
-import org.spongepowered.api.text.format.TextColors;
-import org.spongepowered.api.text.format.TextStyles;
+import com.mojang.brigadier.Command;
+import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import com.mojang.brigadier.context.CommandContext;
+import com.mojang.brigadier.exceptions.CommandSyntaxException;
+import net.minecraft.command.CommandSource;
+import net.minecraft.command.Commands;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.text.TextFormatting;
 
-import java.util.HashMap;
-import java.util.Optional;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
+import static com.hiroku.tournaments.util.CommandUtils.getOptArgument;
 
 /**
  * Command for checking, saving, loading, and deleting presets.
  *
  * @author Hiroku
  */
-public class PresetsCommand implements CommandExecutor {
-	public static CommandSpec getSpec() {
-		HashMap<String, String> choices = new HashMap<>();
-		choices.put("save", "save");
-		choices.put("load", "load");
-		choices.put("delete", "delete");
-		choices.put("rename", "rename");
-		choices.put("", "check");
+public class PresetsCommand implements Command<CommandSource> {
+	Set<String> CHOICES = new HashSet<>(Arrays.asList("save", "load", "delete", "rename", ""));
 
-		return CommandSpec.builder()
-				.permission("tournaments.command.admin.presets")
-				.description(Text.of("Adds, removes, or checks rule preset details"))
-				.arguments(
-						GenericArguments.optional(GenericArguments.choices(Text.of("option"), choices)),
-						GenericArguments.optional(GenericArguments.remainingJoinedStrings(Text.of("preset-name"))))
-				.executor(new PresetsCommand())
-				.build();
+	public LiteralArgumentBuilder<CommandSource> create() {
+		return Commands.literal("presets")
+//				.description(Text.of("Adds, removes, or checks rule preset details"))
+				.requires(source -> User.hasPermission(source, "tournaments.command.admin.presets"))
+				.executes(this)
+				.then(
+						Commands.argument("choice", ChoiceSetArgument.choiceSet(CHOICES))
+								.executes(this)
+								.then(
+										Commands.argument("preset", StringArgumentType.greedyString())
+												.executes(this)
+								)
+				);
 	}
 
 	@Override
-	public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
-		Optional<String> optChoice = args.<String>getOne(Text.of("option"));
-		String choice = "check";
-		if (optChoice.isPresent())
-			choice = optChoice.get();
-		Optional<String> optPresetName = args.<String>getOne(Text.of("preset-name"));
-		if (!choice.equals("check") && !optPresetName.isPresent()) {
-			src.sendMessage(Text.of(TextColors.RED, "Missing argument: preset-name"));
-			return CommandResult.empty();
+	public int run(CommandContext<CommandSource> context) throws CommandSyntaxException {
+		String choice = getOptArgument(context, "choice", String.class).orElse("check");
+		String presetName = getOptArgument(context, "preset", String.class).orElse(null);
+		if (!choice.equals("check") && presetName == null) {
+			context.getSource().sendFeedback(Text.of(TextFormatting.RED, "Missing argument: preset-name"), true);
+			return 0;
 		}
 
 		if (choice.equals("check")) {
-			boolean isPlayer = src instanceof Player;
+			boolean isPlayer = context.getSource().getEntity() instanceof PlayerEntity;
 			ImmutableMap<String, Preset> presets = Presets.getPresets();
-			src.sendMessage(Text.of(TextColors.GOLD, "------- Rule Presets --------"));
-			src.sendMessage(Text.of(TextColors.GRAY, TextStyles.ITALIC, "Hover to see preset details"));
+			context.getSource().sendFeedback(Text.of(TextFormatting.GOLD, "------- Rule Presets --------"), true);
+			context.getSource().sendFeedback(Text.of(TextFormatting.GRAY, TextFormatting.ITALIC, "Hover to see preset details"), true);
 			for (String name : presets.keySet()) {
 				Text.Builder builder = Text.builder();
-				if (isPlayer) {
-					builder.append(Text.of(TextColors.RED, TextActions.executeCallback(dummySrc ->
-					{
-						if (Presets.getPreset(name) != null) {
-							Presets.deletePreset(name);
-							src.sendMessage(Text.of(TextColors.GRAY, "Successfully deleted preset: ", TextColors.DARK_AQUA, name));
-						}
-					}), TextActions.showText(Text.of(TextColors.GRAY, TextStyles.ITALIC, "Click to delete this preset")), "[Delete]"));
-				}
-				builder.append(Text.of(TextColors.GOLD, "  ", TextColors.DARK_AQUA, TextActions.showText(presets.get(name).getDisplayText()), name));
-				src.sendMessage(builder.build());
+				// TODO textactions
+//				if (isPlayer) {
+//					builder.append(Text.of(TextFormatting.RED, TextActions.executeCallback(dummySrc ->
+//					{
+//						if (Presets.getPreset(name) != null) {
+//							Presets.deletePreset(name);
+//							context.getSource().sendFeedback(Text.of(TextFormatting.GRAY, "Successfully deleted preset: ", TextFormatting.DARK_AQUA, name), true);
+//						}
+//					}), TextActions.showText(Text.of(TextFormatting.GRAY, TextFormatting.ITALIC, "Click to delete this preset")), "[Delete]"));
+//				}
+//				builder.append(Text.of(TextFormatting.GOLD, "  ", TextFormatting.DARK_AQUA, TextActions.showText(presets.get(name).getDisplayText()), name));
+				context.getSource().sendFeedback(builder.build(), true);
 			}
 			if (presets.isEmpty())
-				src.sendMessage(Text.of(TextColors.RED, "None."));
-			return CommandResult.success();
-		}
-		String arg = optPresetName.get();
-		if (choice.equals("rename")) {
-			String[] splits = arg.split(" ");
+				context.getSource().sendFeedback(Text.of(TextFormatting.RED, "None."), true);
+			return 1;
+		} else if (choice.equals("rename")) {
+			String[] splits = presetName.split(" ");
 			if (splits.length < 2) {
-				src.sendMessage(Text.of(TextColors.RED, "Not enough arguments: /tournaments presets rename <oldName> <newName>"));
-				return CommandResult.empty();
+				context.getSource().sendFeedback(Text.of(TextFormatting.RED, "Not enough arguments: /tournaments presets rename <oldName> <newName>"), true);
+				return 0;
 			}
 			String oldName = Presets.getMatchingKey(splits[0]);
 			String newName = splits[1];
 			if (oldName == null) {
-				src.sendMessage(Text.of(TextColors.RED, "Unknown preset: ", splits[0]));
-				return CommandResult.empty();
+				context.getSource().sendFeedback(Text.of(TextFormatting.RED, "Unknown preset: ", splits[0]), true);
+				return 0;
 			}
 			Presets.renamePreset(oldName, newName);
-			src.sendMessage(Text.of(TextColors.DARK_GREEN, "Renamed ", TextColors.DARK_AQUA, oldName, TextColors.DARK_GREEN, " to ", TextColors.DARK_AQUA, newName));
-			return CommandResult.success();
+			context.getSource().sendFeedback(Text.of(TextFormatting.DARK_GREEN, "Renamed ", TextFormatting.DARK_AQUA, oldName, TextFormatting.DARK_GREEN, " to ", TextFormatting.DARK_AQUA, newName), true);
+			return 1;
 		} else {
 			if (Tournament.instance() == null) {
-				src.sendMessage(Text.of(TextColors.RED, "No tournament. Try /tournament create"));
-				return CommandResult.empty();
+				context.getSource().sendFeedback(Text.of(TextFormatting.RED, "No tournament. Try /tournament create"), true);
+				return 0;
 			}
+
 			if (choice.equals("save")) {
 				Preset preset = new Preset(Tournament.instance().getRuleSet(),
 						Tournament.instance().rewards,
 						Zones.INSTANCE.getEngagedZones());
-				Presets.setPreset(arg, preset);
-				src.sendMessage(Text.of(TextColors.DARK_GREEN, "Successfully set preset ", Text.of(TextColors.DARK_AQUA, TextActions.showText(preset.getDisplayText()), arg)));
-				return CommandResult.success();
+				Presets.setPreset(presetName, preset);
+				// TODO textactions
+//				context.getSource().sendFeedback(Text.of(TextFormatting.DARK_GREEN, "Successfully set preset ", Text.of(TextFormatting.DARK_AQUA, TextActions.showText(preset.getDisplayText()), presetName)));
+				return 1;
 			}
 
-			String name = Presets.getMatchingKey(arg);
+			String name = Presets.getMatchingKey(presetName);
 			if (name == null) {
-				src.sendMessage(Text.of(TextColors.RED, "Unknown preset: ", TextColors.DARK_AQUA, arg));
-				return CommandResult.empty();
+				context.getSource().sendFeedback(Text.of(TextFormatting.RED, "Unknown preset: ", TextFormatting.DARK_AQUA, presetName), true);
+				return 0;
 			}
 
 			if (choice.equals("load")) {
@@ -125,14 +125,15 @@ public class PresetsCommand implements CommandExecutor {
 					for (Zone zone : Zones.INSTANCE.getZones())
 						if (!preset.zones.contains(zone))
 							zone.engaged = false;
-				src.sendMessage(Text.of(TextColors.DARK_GREEN, "Successfully loaded preset: ", TextColors.DARK_AQUA, name));
-				return CommandResult.success();
-			} else // "delete"
-			{
+				context.getSource().sendFeedback(Text.of(TextFormatting.DARK_GREEN, "Successfully loaded preset: ", TextFormatting.DARK_AQUA, name), true);
+				return 1;
+			} else if (choice.equals("delete")) {
 				Presets.deletePreset(name);
-				src.sendMessage(Text.of(TextColors.GRAY, "Successfully deleted preset: ", TextColors.DARK_AQUA, name));
-				return CommandResult.success();
+				context.getSource().sendFeedback(Text.of(TextFormatting.GRAY, "Successfully deleted preset: ", TextFormatting.DARK_AQUA, name), true);
+				return 1;
 			}
 		}
+
+		return 0;
 	}
 }
